network: caching: move interface defintions
Move the following interface definitions into the `Caching.Abstract'
namespace:

	* ISensorCache;
	* IDataCache.
