+++ SensateIoT.Platform/SensateIoT.Platform.Network/SensateIoT.Platform.Network.Common/Caching/Abstract/IDataCache.cs
@@ -17,7 +17,7 @@ using SensateIoT.Platform.Network.Data.Models;
 using ApiKey = SensateIoT.Platform.Network.Data.DTO.ApiKey;
 using Sensor = SensateIoT.Platform.Network.Data.DTO.Sensor;
 
-namespace SensateIoT.Platform.Network.Common.Caching.Object
+namespace SensateIoT.Platform.Network.Common.Caching.Abstract
 {
 	public interface IDataCache : IDisposable
 	{
